bruh

+++ src/server/network/protocol/protocolstatus.cpp
@@ -188,7 +188,7 @@ void ProtocolStatus::sendInfo(uint16_t requestedInfo, const std::string& charact
 
 	if (requestedInfo & REQUEST_PLAYERS_INFO) {
 		output->addByte(0x20);
-		output->add<uint32_t>(static_cast<size_t>(g_game().getPlayersOnline()));
+		output->add<uint32_t>(static_cast<uint32_t>(g_game().getPlayersOnline()));
 		output->add<uint32_t>(g_configManager().getNumber(MAX_PLAYERS));
 		output->add<uint32_t>(g_game().getPlayersRecord());
 	}
